requirements.txt
aiogram==3.4.1
beautifulsoup4==4.12.2
requests==2.31.0
APScheduler==3.10.4
lxml==4.9.3
python-dotenv==1.0.0
Pillow==10.1.0 